pyproject.toml
[build-system]
requires = ["flit_core ==3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "nessie"
authors = [{name = "Jan-Christoph Klie", email = "git@mrklie.com"}]
readme = "README.md"
license = {file = "LICENSE"}
dynamic = ["version"]
description = "Annotation error detection and correction"
requires-python = ">=3.8.10"
classifiers = ["Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "torch==1.11.0",
    "flair==0.11.*",
    "sacremoses==0.0.53",
    "slurmee",
    "transformers==4.16.1",
    "numpy==1.22.3",
    "scipy==1.7.3",
    "backoff==1.11.1",
    "datasets==1.18.2",
    "pandas==1.4.0",
    "wget==3.2",
    "pyro-ppl==1.8.0",
    "cleanlab==1.0",
    "lightgbm==3.3.2",
    "diskcache==5.4.0",
    "fasttext==0.9.2",
    "cleantext==1.1.4",
    "sentence-transformers==2.1.0",
    "sklearn-crfsuite==0.3.6",
    "seqeval==1.2.2",
    "faiss-cpu==1.7.2",
    "crowd-kit==0.0.9",
    "iobes==1.5.1",
    "suffix-tree==0.0.7",
    "awkward==1.7.0",
    "netcal==1.2.1",
    "pooch==1.6.0",
    "ireval==0.1.1",
    "jupyter==1.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest==6.2",
    "black==22.1.0",
    "isort==5.10.1"
]

doc = [
    "seaborn==0.11.2",
    "notebook==6.4.8",
    "Sphinx==4.4.0",
    "m2r2==0.3.2",
    "sphinx-autodoc-typehints==1.17.0",
    "sphinx-rtd-theme==1.0.0",
    "sphinx-autoapi==1.8.4",
    "nbsphinx==0.8.8",
    "sphinx-copybutton==0.5.0",
    "nbsphinx-link==1.3.0",
    "nbstripout==0.5.0",
    "pre-commit==2.17.0",
    "sphinx_rst_builder",
]

[project.urls]
Home = "https://github.com/jcklie/nessie/"

[tool.black]
line-length = 120
target-version = ['py38']
include = '\.pyi?$'

